refactor: move `Client` impl blocks out of `credential` module [WPB-15530]
