+++ crypto/src/mls/client/identifier.rs
@@ -50,8 +50,7 @@ impl ClientIdentifier {
             ClientIdentifier::Basic(id) => signature_schemes.iter().try_fold(
                 Vec::with_capacity(signature_schemes.len()),
                 |mut acc, &sc| -> Result<_> {
-                    let cb = Client::new_basic_credential_bundle(&id, sc, backend)
-                        .map_err(RecursiveError::mls_credential("creating new basic credential bundle"))?;
+                    let cb = Client::new_basic_credential_bundle(&id, sc, backend)?;
                     acc.push((sc, id.clone(), cb));
                     Ok(acc)
                 },
@@ -64,8 +63,7 @@ impl ClientIdentifier {
                         let id = cert
                             .get_client_id()
                             .map_err(RecursiveError::mls_credential("getting client id"))?;
-                        let cb = Client::new_x509_credential_bundle(cert)
-                            .map_err(RecursiveError::mls_credential("creating new x509 credential bundle"))?;
+                        let cb = Client::new_x509_credential_bundle(cert)?;
                         acc.push((sc, id, cb));
                         Ok(acc)
                     })
